python-backend/requirements.txt
# ML Dependencies
scikit-learn==1.1.3
pandas==2.1.4
numpy==1.23.5
xgboost==2.0.3
matplotlib==3.5.3
seaborn==0.13.0
joblib==1.2.0

# Rasa Dependencies
rasa==3.6.13
rasa-sdk==3.6.2

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
pydantic==1.10.9
httpx==0.25.2
